created the user helper functions
